refactored, added settings, one c file
