+++ sw/airborne/modules/sensors/mag_qmc5883_regs.h
@@ -20,16 +20,16 @@
  */
 
 /**
- * @file peripherals/qmc5883_regs.h
+ * @file modules/sensors/mag_qmc5883_regs.h
  * Register defenitions for the QMC5883 magnetometer.
  */
 
 #ifndef QMC5883_REGS_H
 #define QMC5883_REGS_H
 
-/* default I2C address */
-// #define QMC5883_ADDR 0x1B
-#define QMC5883_ADDR 0x0D
+/* read and write I2C addresses based on default 0x0D */
+#define QMC5883_READ_ADDR  0x1B
+#define QMC5883_WRITE_ADDR 0x1A
 
 /* Registers Axis X,Y,Z */
 #define QMC5883_REG_DATXL  0x00
@@ -43,26 +43,37 @@
 #define QMC5883_REG_STATUS 0x06
 
 /*  Registers Temperature */
-
-/*
-16 bits temperature sensor output is in 2’s complement.
-
-Temperature sensor gain is factory-calibrated, 
-but its offset has not been compensated,
-only relative temperature value is accurate. 
-
-The temperature coefficient is about 100 LSB/°C
-*/
-
-//#define QMC5883_REG_TEMPM  0x07  /* Not so useful ATM, therefore not implemented */
-//#define QMC5883_REG_TEMPL  0x08  /* Not so useful ATM, therefore not implemented */
+// #define QMC5883_REG_TEMPM  0x07  /* Not so useful ATM, therefore not implemented */
+// #define QMC5883_REG_TEMPL  0x08  /* Not so useful ATM, therefore not implemented */
 
 /* Registers Config */
-#define QMC5883_REG_CONTROL_1    0x09 
-#define QMC5883_REG_CONTROL_2    0x0A
+#define QMC5883_REG_CONTROL_1    0x09  /* settings for MODE */
+#define QMC5883_REG_CONTROL_2    0x0A  /* settings for INT_ENB */
 #define QMC5883_REG_RESET_PERIOD 0x0B
 
 //#define QMC5883_REG_IDC    0x0C  /* reserved */
 //#define QMC5883_REG_IDD    0x0D  /* reserved */
+
+/* Settings for CONTROL_1 */
+
+// MODE CONTROL (MODE)
+#define QMC5883_MODE_STBY 0x00
+#define QMC5883_MODE_CONT 0x01
+
+// OUTPUT DATA RATE IN HZ (ODR)
+#define QMC5883_ODR_10  0x00
+#define QMC5883_ODR_50  0x04
+#define QMC5883_ODR_100 0x08
+#define QMC5883_ODR_200 0x0C
+
+// FULL SCALE (RNG)
+#define QMC5883_RNG_2G 0x00
+#define QMC5883_RNG_8G 0x10
+
+// OVER SAMPLE RATIO (OSR)
+#define QMC5883_OSR_512 0x00
+#define QMC5883_OSR_256 0x40
+#define QMC5883_OSR_128 0x80
+#define QMC5883_OSR_64  0xC0
 
 #endif // QMC5883_REGS_H
